Move rules to their own package
Adds a couple of tests, including Mocking objects in some places
